requirements.txt
pandas
numpy
scikit-learn
xgboost
streamlit
pyyaml
joblib
